requirements.txt
# Python dependencies for AIResQ AppDeploy
python-dotenv>=0.19.0
